requirements.txt
streamlit
pdfplumber
numpy
faiss-cpu
sentence-transformers==2.2.2
transformers==4.49.0
langchain
langchain-groq